backtest_pipeline重写&并集

+++ src/data_loader.py
@@ -4,10 +4,23 @@ from config import params
 
 
 def get_daily_price_pd(usecols=[]):
+    file_dir = './data/processed/merge_data.parquet'
     if len(usecols) == 0:
-        df = pd.read_parquet('./data/processed/merge_data.parquet')
+        df = pd.read_parquet(file_dir)
     else:
-        df = pd.read_parquet('./data/processed/merge_data.parquet', columns=usecols)
+        df = pd.read_parquet(file_dir, columns=usecols)
+    # df = pd.read_csv(os.path.join(params['data_dir'], 'merge_data_ret.csv'), encoding='utf-8-sig', usecols=usecols)
+    # if usecols:
+    #     df = df[usecols]
+    return df
+
+
+def get_daily_price_ret_pd(usecols=[]):
+    file_dir = './data/processed/merge_data_ret.parquet'
+    if len(usecols) == 0:
+        df = pd.read_parquet(file_dir)
+    else:
+        df = pd.read_parquet(file_dir, columns=usecols)
     # df = pd.read_csv(os.path.join(params['data_dir'], 'merge_data_ret.csv'), encoding='utf-8-sig', usecols=usecols)
     # if usecols:
     #     df = df[usecols]
@@ -34,6 +47,20 @@ def csv_2_parquet(file_path=""):
         df = pd.read_csv(file_path, encoding='utf-8-sig', low_memory=False)
         df = df.astype({col: str for col in df.select_dtypes(include=['object']).columns})
         parquet_path = file_path.replace('.csv', '.parquet')
+        df.to_parquet(parquet_path, engine='pyarrow', index=False)
+        print(f"转换成功: {file_path} -> {parquet_path}")
+    except Exception as e:
+        print(f"转换失败: {file_path}, 错误信息: {e}")
+
+def dta_2_parquet(file_path=""):
+    if not os.path.isfile(file_path) or not file_path.endswith('.dta'):
+        print(f"无效的 CSV 文件路径: {file_path}")
+        return
+
+    try:
+        df = pd.read_stata(file_path)
+        df = df.astype({col: str for col in df.select_dtypes(include=['object']).columns})
+        parquet_path = file_path.replace('.dta', '.parquet')
         df.to_parquet(parquet_path, engine='pyarrow', index=False)
         print(f"转换成功: {file_path} -> {parquet_path}")
     except Exception as e:
@@ -66,4 +93,5 @@ if __name__ == '__main__':
     # csv_2_parquet('../data/raw/season_500_0512.csv')
     # csv_2_parquet(file_path='../data/raw/merge_final.csv')
     # par_code_2_str(parquet_path='../data/raw/merge_final.parquet', cols=['证券代码'])
-    par_code_2_str(parquet_path='../data/raw/season_500_0512.parquet', cols=['证券代码'])
+    # par_code_2_str(parquet_path='../data/raw/season_500_0512.parquet', cols=['证券代码'])
+    # dta_2_parquet(file_path='../data/raw/merge_final.dta')
